Enable SDIO and add missing WIFI nodes
These edits will enable the wifi combo chip on the rockpro for the 2.0 AND 2.1 boards. There is still firmware required for it to work that RFKILL will need to load, but that's outside the scope of this PR.

NOTE: There is a possibility this change will cause a conflict with the PCIE being enabled. I have not disabled it in this PR, so if you encounter instability then I recommend disabling pcie0 and pcie_phy.